requirements.txt
asgiref==3.5.2
black==22.6.0
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-import-export==2.8.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
mypy-extensions==0.4.3
mysqlclient==2.1.1
nepali-datetime==1.0.7
odfpy==1.4.1
openpyxl==3.0.10
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.10
webencodings==0.5.1
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
gunicorn==20.1.0